backend/requirements.txt
Django==2.2.28
django-cors-headers==3.7.0
djangorestframework==3.11.2
djangorestframework-simplejwt==4.6.0
psycopg2==2.9.2
PyJWT==2.10.1
pytz==2025.1
sqlparse==0.5.3
